add router & change API result
