+++ LLM_verify_model.py
@@ -40,6 +40,7 @@ class ImageVerifyModel :
             "분위기가 아니라 물체가 존재해야합니다. 텀블러를 사용한 것이 맞으면 모두 '예'로 출력해주세요."
             "적합한 이미지인지 예/아니오로 대답해주세요."
             "대답에 따른 이유도 간단히 한 줄로 설명해주세요. (30자 넘지 않게)"
+            "판단이 '예'인 경우에는 이유는 생략하고, '예'만 출력해주세요."
         )
         # API 호출
         result = self.model.generate_content(
