+++ worker.py
@@ -30,6 +30,10 @@ def run_worker():
             # 로깅용
             print(f"인증 결과: {result}")
 
+            # '예' 여부가 정확히 일치할 때만 True
+            # is_verified = result.strip() == "예"
+            is_verified = "예" in result.strip().lower()
+
             # 콜백 URL 내 challengeId 치환
             # -> CALLBACK_URL에 {verificationId}가 포함되는 경우, Python에서 실제 전송 전에 .format() 또는 f-string으로 치환해줘야함 
             formatted_url = os.getenv("CALLBACK_URL").format(verificationId=data["verificationId"])
@@ -41,7 +45,7 @@ def run_worker():
                 "memberId": data["memberId"],
                 "challengeId": data["challengeId"],
                 "date": data["date"],
-                "result": result
+                "result": is_verified
             })
 
             message.ack()
